requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.2
Werkzeug==0.10.1
argparse==1.2.1
itsdangerous==0.24
py==1.4.26
pycrypto==2.6.1
pytest==2.6.4
wsgiref==0.1.2
